+++ collector/types.go
@@ -23,6 +23,7 @@ type Exporter struct {
 	totalScrapes        prometheus.Counter
 	scrapeErrors        *prometheus.CounterVec
 	scrapeResults       []prometheus.Metric
+	scrapeRequests      chan struct{}
 	databases           []*Database
 	logger              *slog.Logger
 	allConstLabels      []string
@@ -64,23 +65,7 @@ type MetricCacheRecord struct {
 }
 
 type Config struct {
-	ConfigFile         string
-	User               string
-	Password           string
-	ConnectString      string
-	DbRole             string
-	ConfigDir          string
-	ExternalAuth       bool
-	MaxIdleConns       int
-	MaxOpenConns       int
-	PoolIncrement      int
-	PoolMaxConnections int
-	PoolMinConnections int
-	CustomMetrics      string
-	QueryTimeout       int
-	DefaultMetricsFile string
-	ScrapeInterval     time.Duration
-	LoggingConfig      LoggingConfig
+	ConfigFile string
 }
 
 // Metric is an object description